squash merge feature/eventbridge_v2/ecs-target
