+++ tests/aws/services/events/conftest.py
@@ -11,72 +11,7 @@ from tests.aws.services.events.helper_functions import put_entries_assert_result
 
 LOG = logging.getLogger(__name__)
 
-
-@pytest.fixture
-def events_create_event_bus(aws_client, region_name, account_id):
-    event_bus_names = []
-
-    def _create_event_bus(**kwargs):
-        if "Name" not in kwargs:
-            kwargs["Name"] = f"test-event-bus-{short_uid()}"
-
-        response = aws_client.events.create_event_bus(**kwargs)
-        event_bus_names.append(kwargs["Name"])
-        return response
-
-    yield _create_event_bus
-
-    for event_bus_name in event_bus_names:
-        try:
-            response = aws_client.events.list_rules(EventBusName=event_bus_name)
-            rules = [rule["Name"] for rule in response["Rules"]]
-
-            # Delete all rules for the current event bus
-            for rule in rules:
-                try:
-                    response = aws_client.events.list_targets_by_rule(
-                        Rule=rule, EventBusName=event_bus_name
-                    )
-                    targets = [target["Id"] for target in response["Targets"]]
-
-                    # Remove all targets for the current rule
-                    if targets:
-                        for target in targets:
-                            aws_client.events.remove_targets(
-                                Rule=rule, EventBusName=event_bus_name, Ids=[target]
-                            )
-
-                    aws_client.events.delete_rule(Name=rule, EventBusName=event_bus_name)
-                except Exception as e:
-                    LOG.warning(
-                        "Failed to delete rule %s: %s",
-                        rule,
-                        e,
-                    )
-
-            # Delete archives for event bus
-            event_source_arn = (
-                f"arn:aws:events:{region_name}:{account_id}:event-bus/{event_bus_name}"
-            )
-            response = aws_client.events.list_archives(EventSourceArn=event_source_arn)
-            archives = [archive["ArchiveName"] for archive in response["Archives"]]
-            for archive in archives:
-                try:
-                    aws_client.events.delete_archive(ArchiveName=archive)
-                except Exception as e:
-                    LOG.warning(
-                        "Failed to delete archive %s: %s",
-                        archive,
-                        e,
-                    )
-
-            aws_client.events.delete_event_bus(Name=event_bus_name)
-        except Exception as e:
-            LOG.warning(
-                "Failed to delete event bus %s: %s",
-                event_bus_name,
-                e,
-            )
+# some fixtures are shared in localstack/testing/pytest/fixtures.py
 
 
 @pytest.fixture
@@ -128,43 +63,6 @@ def create_role_event_bus_source_to_bus_target(create_iam_role_with_policy):
         return role_arn_bus_source_to_bus_target
 
     yield _create_role_event_bus_to_bus
-
-
-@pytest.fixture
-def events_put_rule(aws_client):
-    rules = []
-
-    def _put_rule(**kwargs):
-        if "Name" not in kwargs:
-            kwargs["Name"] = f"rule-{short_uid()}"
-
-        response = aws_client.events.put_rule(**kwargs)
-        rules.append((kwargs["Name"], kwargs.get("EventBusName", "default")))
-        return response
-
-    yield _put_rule
-
-    for rule, event_bus_name in rules:
-        try:
-            response = aws_client.events.list_targets_by_rule(
-                Rule=rule, EventBusName=event_bus_name
-            )
-            targets = [target["Id"] for target in response["Targets"]]
-
-            # Remove all targets for the current rule
-            if targets:
-                for target in targets:
-                    aws_client.events.remove_targets(
-                        Rule=rule, EventBusName=event_bus_name, Ids=[target]
-                    )
-
-            aws_client.events.delete_rule(Name=rule, EventBusName=event_bus_name)
-        except Exception as e:
-            LOG.warning(
-                "Failed to delete rule %s: %s",
-                rule,
-                e,
-            )
 
 
 @pytest.fixture
@@ -261,44 +159,6 @@ def put_event_to_archive(aws_client, events_create_event_bus, events_create_arch
 
 
 @pytest.fixture
-def create_sqs_events_target(aws_client, sqs_get_queue_arn):
-    queue_urls = []
-
-    def _create_sqs_events_target(queue_name: str | None = None) -> tuple[str, str]:
-        if not queue_name:
-            queue_name = f"tests-queue-{short_uid()}"
-        sqs_client = aws_client.sqs
-        queue_url = sqs_client.create_queue(QueueName=queue_name)["QueueUrl"]
-        queue_urls.append(queue_url)
-        queue_arn = sqs_get_queue_arn(queue_url)
-        policy = {
-            "Version": "2012-10-17",
-            "Id": f"sqs-eventbridge-{short_uid()}",
-            "Statement": [
-                {
-                    "Sid": f"SendMessage-{short_uid()}",
-                    "Effect": "Allow",
-                    "Principal": {"Service": "events.amazonaws.com"},
-                    "Action": "sqs:SendMessage",
-                    "Resource": queue_arn,
-                }
-            ],
-        }
-        sqs_client.set_queue_attributes(
-            QueueUrl=queue_url, Attributes={"Policy": json.dumps(policy)}
-        )
-        return queue_url, queue_arn
-
-    yield _create_sqs_events_target
-
-    for queue_url in queue_urls:
-        try:
-            aws_client.sqs.delete_queue(QueueUrl=queue_url)
-        except Exception as e:
-            LOG.debug("error cleaning up queue %s: %s", queue_url, e)
-
-
-@pytest.fixture
 def events_allow_event_rule_to_sqs_queue(aws_client):
     def _allow_event_rule(sqs_queue_url, sqs_queue_arn, event_rule_arn) -> None:
         # allow event rule to write to sqs queue
@@ -327,7 +187,7 @@ def events_allow_event_rule_to_sqs_queue(aws_client):
 
 @pytest.fixture
 def put_events_with_filter_to_sqs(
-    aws_client, events_create_event_bus, events_put_rule, create_sqs_events_target
+    aws_client, events_create_event_bus, events_put_rule, sqs_as_events_target
 ):
     def _put_events_with_filter_to_sqs(
         pattern: dict,
@@ -342,7 +202,7 @@ def put_events_with_filter_to_sqs(
             event_bus_name = f"test-bus-{short_uid()}"
             events_create_event_bus(Name=event_bus_name)
 
-        queue_url, queue_arn = create_sqs_events_target()
+        queue_url, queue_arn = sqs_as_events_target()
 
         events_put_rule(
             Name=rule_name,
